airflow/requirements.txt
apache-airflow==2.4.3
pandas
psycopg2-binary
pymysql
gspread_pandas
